Add Hikeytv Product to lunch menu
hikeytv.mk product make file

kernel_defconfig is a defconfig for the Linaro Hikey 4.4 Kernel
which adds all the multimedia device drivers such as remote controls
IR, DVB USB Devices etc

vendorsetup.sh contains a couple of cheeky helper functions to help
with a quick turn around of a kernel build to aid with testing.
A little hacky but does the job.